View, BaseView 및 Model 전달 오류 수정
Window를 전달해서 통합적으로 처리하는 것으로 변경

+++ V1/Library/DmiSessionInterface/Manager.Command.cpp
@@ -33,7 +33,7 @@ void SESSION::Manager::Command::AddCommand(SESSION::Command::Root * pcCommand)
 }
 
 // 2. Command Vector 크기를 가져온다.
-int SESSION::Manager::Command::GetCommandSize()
+size_t SESSION::Manager::Command::GetCommandSize()
 {
 	return m_vCommand.size();
 }
